_archive_v2/02_Basic_RAG/requirements.txt
# Core
cocoindex==0.3.25
pgserver==0.1.4

# Vector Store
chromadb==1.4.0

# LangChain
langchain==1.2.3
langchain-community==0.4.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langchain-google-genai==4.1.3
langchain-ollama==1.0.1

# PDF Processing
pypdf==6.6.0

# Utilities
requests==2.32.5
python-dotenv==1.2.1
